miselienv/requirements.txt
asgiref>=3.5.0
Django>=4.0.3
numpy>=1.22.1
pandas>=1.4.1
python-dateutil>=2.8.2
pytz>=2021.3
six>=1.16.0
psycopg2>=2.9.3
sqlparse>=0.4.2
tzdata>=2021.5
setuptools>=60.5.0
platformdirs>=2.5.1
filelock>=3.4.2
pip>=22.0.3
pip-review>=1.1.1
virtualenv>=20.13.1
virtualenv-clone>=0.5.7
django-tenants>=3.4.2
django-import-export>=2.7.1
django-cors-headers>=3.11.0
djangorestframework>=3.13.1
pytest>=7.1.1